\name{skyspline-package}
\alias{skyspline-package}
\alias{skyspline}
\docType{package}
\title{
  Phylodynamic inference using birth-rate splines. 
}
\description{
	This package includes functions for MLE/bootstrap and Bayesian MCMC estimation of population size and birth rate through time given genealogical data. 
}
\details{
  Use fit.skyspline.ml for maximum likelihood or maximum a posterior inference. 
  parboot.skyspline.mle can be used to estimate CIs. 
  skyspline.metrop.hastings provides a simple Metropolis-Hastings sampler for Bayesian inference. 
}
\author{
Erik M Volz

Maintainer: Erik Volz <erik.volz@gmail.com>
}
\references{
##
}
\keyword{ phylodynamics }
\seealso{
}
\examples{
  \dontrun{
  }
}
